Skip initializing test env in evergreen package if it's doing the auth test
